PDF_Chat_Bot/requirements.txt
streamlit
pypdf2
langchain
python-dotenv
openai
huggingface_hub
InstructorEmbedding
sentence_transformers
faiss-cpu

#-e .